implemented phase 5, updates on docs

+++ src/main/kotlin/com/solodev/fleet/modules/accounts/application/dto/AccountBalanceResponse.kt
@@ -2,4 +2,4 @@ package com.solodev.fleet.modules.accounts.application.dto
 
 import kotlinx.serialization.Serializable
 
-@Serializable data class AccountBalanceResponse(val account: String, val balance: Double)
+@Serializable data class AccountBalanceResponse(val account: String, val balance: Long)
