+++ src/main/kotlin/com/solodev/fleet/modules/accounts/application/dto/PaymentRequest.kt
@@ -3,4 +3,4 @@ package com.solodev.fleet.modules.accounts.application.dto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class PaymentRequest(val amount: Double, val paymentMethod: String, val notes: String? = null)
+data class PaymentRequest(val amount: Int, val paymentMethod: String, val notes: String? = null)
